+++ app/mobile/lib/core/i18n/strings.g.dart
@@ -4,9 +4,9 @@
 /// To regenerate, run: `dart run slang`
 ///
 /// Locales: 3
-/// Strings: 12960 (4320 per locale)
+/// Strings: 13005 (4335 per locale)
 ///
-/// Built on 2026-08-07 at 08:42 UTC
+/// Built on 2026-08-07 at 11:28 UTC
 
 // coverage:ignore-file
 // ignore_for_file: type=lint, unused_import